move households to download
